feat: add property set handler

+++ bridge/core/css/computed_css_style_declaration.h
@@ -5,7 +5,6 @@
 #ifndef WEBF_CORE_CSS_COMPUTED_CSS_STYLE_DECLARATION_H_
 #define WEBF_CORE_CSS_COMPUTED_CSS_STYLE_DECLARATION_H_
 
-#include "bindings/qjs/cppgc/member.h"
 #include "core/binding_object.h"
 #include "css_style_declaration.h"
 
@@ -25,7 +24,7 @@ class ComputedCssStyleDeclaration : public CSSStyleDeclaration {
   ScriptValue item(const AtomicString& key, ExceptionState& exception_state) override;
   bool SetItem(const AtomicString& key, const ScriptValue& value, ExceptionState& exception_state) override;
   bool DeleteItem(const webf::AtomicString& key, webf::ExceptionState& exception_state) override;
-  int64_t length() const override;
+  unsigned length() const override;
 
   AtomicString getPropertyValue(const AtomicString& key, ExceptionState& exception_state) override;
   void setProperty(const AtomicString& key, const ScriptValue& value, ExceptionState& exception_state) override;
